Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
psycopg2
pydantic
